[MSELoss] print train_loss after each batch [Example][DLRM] several bug fixes

+++ include/cuda_helper.h
@@ -78,7 +78,7 @@ void print_tensor(const T* ptr, Rect<DIM> rect, const char* prefix)
   int idx = 0;
   printf("%s", prefix);
   for (PointInRectIterator<DIM> it(rect); it(); it++, idx++) {
-    printf(" %.4lf", host_ptr[idx]);
+    printf(" %.4lf", (float)host_ptr[idx]);
   }
   printf("\n");
   checkCUDA(cudaFreeHost(host_ptr));
